requirements.txt
singer-python==5.9.0
google-api-python-client==1.7.11
oauth2client==4.1.3